Adding some test sketches
led_test PWM-dims an array of pins. This will be used to perform the
proximity/signal-strength-based dimming on the dress.

newsoft_xbee is a quick-and-dirty NewSoftSerial-based test of
communicating with the XBee via the Lilypad Arduino. We tried this with
the built-in Serial Monitor, but were getting SIGSEGV on the
Ubuntu/Linux release of the arduino IDE. No problem - just drop into
minicom, and we can see our "+++" being responded to with a hearty "OK".
All good here.

Next up: trying to get the XBees to communicate with each other. Once we
have that step, we can think about on-dress design.