Adding dBm calculation
and testing that out.

We hit a bizarre snag: the RSSI pin is giving us a stream of zeros, even
though we can see the RSSI LED light up! We also added LED blinking on
both the Lilypad and Fio to show us when pings are transmitted/received
- and they light up correctly. We *also* set ATRP FF, so that there's no
potential timing issue with reading the pin.

The result? Still getting a stream of zeros. We're going to try out API
mode next as a workaround, since you can directly read RSSI from the API
frame...keep posted!